+++ batchflow/models/torch/layers/__init__.py
@@ -6,5 +6,5 @@ from .conv import Conv, ConvTranspose, DepthwiseConv, DepthwiseConvTranspose, Se
 from .pooling import MaxPool, AvgPool, Pool, AdaptiveMaxPool, AdaptiveAvgPool, AdaptivePool, \
 					 GlobalPool, GlobalMaxPool, GlobalAvgPool
 from .resize import IncreaseDim, ReduceDim, Reshape, Interpolate, PixelShuffle, SubPixelConv, \
-					Branch, Upsample, Combine, SEBlock, Crop
+					Upsample, Combine, SEBlock, Crop
 from .pyramid import PyramidPooling, ASPP
